Fix incorrect setup of custom resources
Needed to wrap in Input (which allows assignment of both the primitive and the Output-wrapped primitive, with Input accounting for the fact that the Output might not be immediately available and waiting until it is) but also needed non primitive values wrapped too. The complex types needed to inherit from ResourceArgs to be compatible with Input. The Input annotation was necessary to ensure the type yielded the correct yaml post-Pulumi doing its thing